+++ changelog.md
@@ -1,8 +1,12 @@
 ## v0.5.7
+### StorageManager changes
+- remove `otherModels`, `docTypes`, `documentIds` properties
+- `loadDocuments` takes an optional `status` parameter of type `WalletStorage.DocumentStatus` (default is `issued`)
+- new variable `@Published public private(set) var deferredDocuments: [WalletStorage.Document] = []` (documents that are not yet issued)
+### Other changes
 - Update eudi-lib-ios-openid4vci-swift to version 0.3.2
 - Rename `OfferedIssueModel` to `OfferedIssuanceModel` 
 - `EudiWallet`: added property `public var accessGroup: String?` (used for sharing keychain items between apps with the same access group)
-- StorageManager: remove `otherModels`, `docTypes`, `documentIds` properties
 
 ## v0.5.6
 - Update eudi-lib-ios-siop-openid4vp-swift to version 0.3.2